refactor: extract shared login form
